added pytest refactored signals
