+++ eventkit_cloud/feature_selection/tests/test_feature_selection.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import
 
 import unittest
-from feature_selection.feature_selection import FeatureSelection
+from ..feature_selection import FeatureSelection
 
 ZIP_README = """
 This thematic file was generated by the HOT Exports Tool.
