+++ eventkit_cloud/jobs/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'eventkit_cloud.jobs.apps.EventKitJobs'
