+++ eventkit_cloud/jobs/apps.py
@@ -1,0 +1,11 @@
+from __future__ import absolute_import
+
+from django.apps import AppConfig
+
+
+class EventKitJobs(AppConfig):
+    name = 'eventkit_cloud.jobs'
+    verbose_name = "Eventkit-Cloud Jobs"
+
+    def ready(self):
+        from .signals import *
