+++ eventkit_cloud/jobs/signals.py
@@ -1,0 +1,14 @@
+from django.db.models.signals import post_save
+from django.dispatch.dispatcher import receiver
+from django.contrib.auth.models import Group, User
+
+
+@receiver(post_save, sender=User)
+def user_post_save(sender, instance, created, **kwargs):
+    """
+    This method is executed whenever a User object is created.
+
+    Adds the new user to DefaultExportExtentGroup.
+    """
+    if created:
+        instance.groups.add(Group.objects.get(name='DefaultExportExtentGroup'))
