+++ eventkit_cloud/settings/tests.py
@@ -1,0 +1,11 @@
+from .prod import *  # NOQA
+
+#Override settings here for test purposes.
+
+CELERY_ALWAYS_EAGER = True
+CELERY_EAGER_PROPAGATES_EXCEPTIONS = True
+BROKER_BACKEND = 'memory'
+
+PASSWORD_HASHERS = (
+    'django.contrib.auth.hashers.MD5PasswordHasher',
+)
