+++ eventkit_cloud/tasks/__init__.py
@@ -4,3 +4,5 @@ from __future__ import absolute_import
 from .scheduled_tasks import *
 from .export_tasks import *
 from .debug_tasks import *
+
+default_app_config = 'eventkit_cloud.tasks.apps.EventKitTasks'
