+++ eventkit_cloud/tasks/models.py
@@ -2,20 +2,13 @@
 from __future__ import unicode_literals
 
 import logging
-import shutil
-import os
 
-from django.conf import settings
 from django.contrib.auth.models import User
 from django.db import models
-from django.db.models.signals import pre_delete, post_delete
-from django.dispatch.dispatcher import receiver
 from django.utils import timezone
 
 from ..jobs.models import Job, LowerCaseCharField
-from ..utils.s3 import delete_from_s3
 from ..core.models import UIDMixin, TimeStampedModelMixin, TimeTrackingModelMixin
-import uuid
 
 logger = logging.getLogger(__name__)
 
@@ -53,6 +46,7 @@ class ExportRun(UIDMixin, TimeStampedModelMixin, TimeTrackingModelMixin):
 
     def soft_delete(self, user=None, *args, **kwargs):
         from .export_tasks import cancel_run
+        from .signals import exportrun_delete_exports
         exportrun_delete_exports(self.__class__, self)
         username = None
         if user:
@@ -153,6 +147,7 @@ class FileProducingTaskResult(models.Model):
         return ret
 
     def soft_delete(self, *args, **kwargs):
+        from .signals import exporttaskresult_delete_exports
         exporttaskresult_delete_exports(self.__class__, self)
         self.deleted = True
         self.save()
@@ -180,34 +175,3 @@ class ExportTaskException(TimeStampedModelMixin):
         managed = True
         db_table = 'export_task_exceptions'
 
-
-@receiver(pre_delete, sender=ExportRun)
-def exportrun_delete_exports(sender, instance, *args, **kwargs):
-    """
-    Delete the associated export files when an ExportRun is deleted.
-    """
-    if getattr(settings, 'USE_S3', False):
-        delete_from_s3(run_uid=str(instance.uid))
-    run_dir = '{0}/{1}'.format(settings.EXPORT_DOWNLOAD_ROOT.rstrip('/'), instance.uid)
-    try:
-        shutil.rmtree(run_dir, ignore_errors=True)
-        logger.info("The directory {0} was deleted.".format(run_dir))
-    except OSError:
-        logger.warn("The directory {0} was already moved or doesn't exist.".format(run_dir))
-
-
-@receiver(pre_delete, sender=FileProducingTaskResult)
-def exporttaskresult_delete_exports(sender, instance, *args, **kwargs):
-    """
-    Delete associated files when deleting the FileProducingTaskResult.
-    """
-    # The url should be constructed as [download context, run_uid, filename]
-    if getattr(settings, 'USE_S3', False):
-        delete_from_s3(download_url=instance.download_url)
-    url_parts = instance.download_url.split('/')
-    full_file_download_path = '/'.join([settings.EXPORT_DOWNLOAD_ROOT.rstrip('/'), url_parts[-2], url_parts[-1]])
-    try:
-        os.remove(full_file_download_path)
-        logger.info("The directory {0} was deleted.".format(full_file_download_path))
-    except OSError:
-        logger.warn("The file {0} was already removed or does not exist.".format(full_file_download_path))
