+++ eventkit_cloud/tasks/tests/test_models.py
@@ -84,7 +84,7 @@ class TestExportRun(TestCase):
         runs = job.runs.all()
         self.assertEquals(0, runs.count())
 
-    @patch('eventkit_cloud.tasks.models.exportrun_delete_exports')
+    @patch('eventkit_cloud.tasks.signals.exportrun_delete_exports')
     def test_soft_delete_export_run(self, mock_run_delete_exports):
         job = Job.objects.first()
         run = ExportRun.objects.create(job=job, user=job.user)
@@ -130,7 +130,7 @@ class TestExportTask(TestCase):
         saved_task = ExportTask.objects.get(uid=self.task_uid)
         self.assertEqual(saved_task, self.task)
 
-    @patch('eventkit_cloud.tasks.models.exporttaskresult_delete_exports')
+    @patch('eventkit_cloud.tasks.signals.exporttaskresult_delete_exports')
     def test_export_task_result(self, mock_etr_delete_exports):
         """
         Test FileProducingTaskResult.
@@ -149,7 +149,7 @@ class TestExportTask(TestCase):
         self.assertTrue(task.result.deleted)
         mock_etr_delete_exports.assert_called_once()
 
-    @patch('eventkit_cloud.tasks.models.delete_from_s3')
+    @patch('eventkit_cloud.tasks.signals.delete_from_s3')
     def test_exportrun_delete_exports(self, delete_from_s3):
         job = Job.objects.first()
         run = ExportRun.objects.create(job=job, user=job.user)
@@ -162,7 +162,7 @@ class TestExportTask(TestCase):
             delete_from_s3.assert_called_once_with(run_uid=str(run_uid))
 
     @patch('os.remove')
-    @patch('eventkit_cloud.tasks.models.delete_from_s3')
+    @patch('eventkit_cloud.tasks.signals.delete_from_s3')
     def test_exporttaskresult_delete_exports(self, delete_from_s3, remove):
 
         # setup
