--- eventkit_cloud/utils/tests/files/query.osm
@@ -1,1 +1,0 @@
-<osm>some data</osm>
