+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+python_files = tests.py test_*.py
+DJANGO_SETTINGS_MODULE=eventkit_cloud.settings.prod
