+++ requirements-dev.txt
@@ -6,3 +6,6 @@ flake8==3.3.0
 autopep8==1.3
 flower==0.9.1
 pydevd==1.0.0
+pytest==3.2.5
+pytest-django==3.1.2
+pytest-xdist==1.20.1
